misc/HtmlTestSuite/results/test-0964.xhtml

<?xml version="1.0" encoding="UTF-8"?><html xmlns="http://www.w3.org/1999/xhtml">
<head>
<title>HTML Templates: Template element as a descendant of the body element.</title>
<meta href="mailto:a.semenov@unipro.ru" name="author" title="Aleksei Yu. Semenov"/>
<meta content="Template element can be a descendant of the body element" name="assert"/>
<link href="http://www.w3.org/TR/2013/WD-html-templates-20130214/#template-element" rel="help"/>
<script src="/resources/testharness.js"/>
<script src="/resources/testharnessreport.js"/>
<script src="/html/resources/common.js"/>
</head>
<body>
<div id="log"/>
<script type="text/javascript">

testInIFrame('../resources/template-contents.html', function(ctx) {
    var doc = ctx.iframes[0].contentDocument;

    assert_not_equals(doc.body.querySelector('template'), null,
        'Template element should be a descendant of the body element');

}, 'Template element as a descendant of the body element. Test loading from a file');

</script>
</body>
</html>
